03.LCL활동-산출물/AWS-Resource-List/requirements.txt
boto3 # ==1.17.20
jmespath # ==0.10.0
XlsxWriter # ==1.3.7
python-dotenv
# pprint==0.1
pprintpp
pandas
forex_python # ==1.8
lxml # ==4.9.3
finance-datareader
pandas_datareader #==
yfinance